+++ internal/css_ast/css_decl_table.go
@@ -1,5 +1,7 @@
 package css_ast
 
+import "github.com/evanw/esbuild/internal/helpers"
+
 type D uint16
 
 const (
@@ -639,4 +641,19 @@ var KnownDeclarations = map[string]D{
 	"writing-mode":                DWritingMode,
 	"z-index":                     DZIndex,
 	"zoom":                        DZoom,
+}
+
+var typoDetector *helpers.TypoDetector
+
+func MaybeCorrectDeclarationTypo(text string) (string, bool) {
+	if typoDetector == nil {
+		// Lazily-initialize the typo detector for speed when it's not needed
+		valid := make([]string, 0, len(KnownDeclarations))
+		for key := range KnownDeclarations {
+			valid = append(valid, key)
+		}
+		detector := helpers.MakeTypoDetector(valid)
+		typoDetector = &detector
+	}
+	return typoDetector.MaybeCorrectTypo(text)
 }