+++ internal/js_lexer/js_lexer.go
@@ -914,7 +914,8 @@ func (lexer *Lexer) NextJSXElementChild() {
 						replacement = "{'>'}"
 					}
 					msg := logger.Msg{Kind: logger.Error, Data: lexer.tracker.MsgData(logger.Range{Loc: logger.Loc{Start: int32(lexer.end)}, Len: 1},
-						fmt.Sprintf("The character \"%c\" is not valid inside a JSX element, but can be escaped as %q instead", lexer.codePoint, replacement))}
+						fmt.Sprintf("The character \"%c\" is not valid inside a JSX element", lexer.codePoint)),
+						Notes: []logger.MsgData{{Text: fmt.Sprintf("Did you mean to escape it as %q instead?", replacement)}}}
 					msg.Data.Location.Suggestion = replacement
 					if !lexer.ts.Parse {
 						// TypeScript treats this as an error but Babel doesn't treat this